pyproject.toml
[project]
name = "docs-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.4.247",
    "python-dotenv>=1.0.1",
    "surrealdb>=0.4.1",
]
